Added pg_dump of all tables except version
This was command used to create all the common tables that are currently
in existence.

 $ pg_dump \
        --schema-only \
        --exclude-table=version \
        --no-privileges \
        --no-owner \
        --encoding=utf-8 \
        --serializable-deferrable \
        --host=$(\
                docker inspect \
                        --format '{{ .NetworkSettings.IPAddress }}' \
                        make_postgres_integration_test_1 \
                ) \
        --username=postgres -W \
   > migration/sql-files/001-common.sql

As you can see I've tried to limit the schema as must as possible to
make it universally applicable. Still in a follow up commit I'm going to
have to make some adjustments to this dump before we can apply it.