Make sure the database is populated with the correct types (e.g. system.bug etc.)
It is okay for integration tests to create the database entries that
they need. That is what the migration.Populate() function does.
Previously this function was called Perform() for historical reasons.
The actual structural migration of a database is performed in the
Migrate() function and not in Populate(). Whether or not this should be
outsourced can be discussed. But for now the integration tests that need
some test data to be present in the database are responsible for
initiating the population themselves.

We can refactor the whole concept of resource.Require() in another issue
but not here.